requirements.txt
streamlit==1.28.1
boto3==1.34.0
python-dotenv==1.0.0
botocore==1.34.0